+++ NEWS.md
@@ -1,5 +1,7 @@
 # openmpp (development version)
 
+* We have added new functionality for users to specify their own custom connections to the OpenM++ OMS API. This should make the package more flexible for different types of user authentication.
+
 # openmpp 0.0.1
 
 * Initial public release.